Revert "[Core] Cleanup C++ code (#15109)"
This reverts commit 64cc092959264cea42099690f13eceb6036e34d8.

+++ src/ray/object_manager/object_manager.cc
@@ -70,7 +70,9 @@ ObjectManager::ObjectManager(
                 [this, object_info,
                  add_object_callback = std::move(add_object_callback)]() {
                   HandleObjectAdded(object_info);
-                  add_object_callback(object_info);
+                  if (add_object_callback) {
+                    add_object_callback(object_info);
+                  }
                 },
                 "ObjectManager.ObjectAdded");
           },
@@ -81,7 +83,9 @@ ObjectManager::ObjectManager(
                 [this, object_id,
                  delete_object_callback = std::move(delete_object_callback)]() {
                   HandleObjectDeleted(object_id);
-                  delete_object_callback(object_id);
+                  if (delete_object_callback) {
+                    delete_object_callback(object_id);
+                  }
                 },
                 "ObjectManager.ObjectDeleted");
           }),
@@ -127,7 +131,10 @@ ObjectManager::ObjectManager(
       [spill_objects_callback, object_store_full_callback]() {
         // TODO(swang): This copies the out-of-memory handling in the
         // CreateRequestQueue. It would be nice to unify these.
-        object_store_full_callback();
+        if (object_store_full_callback) {
+          object_store_full_callback();
+        }
+
         static_cast<void>(spill_objects_callback());
       }));
   // Start object manager rpc server and send & receive request threads
@@ -136,10 +143,17 @@ ObjectManager::ObjectManager(
 
 ObjectManager::~ObjectManager() { StopRpcService(); }
 
-void ObjectManager::Stop() { plasma::plasma_store_runner->Stop(); }
+void ObjectManager::Stop() {
+  if (plasma::plasma_store_runner != nullptr) {
+    plasma::plasma_store_runner->Stop();
+  }
+}
 
 bool ObjectManager::IsPlasmaObjectSpillable(const ObjectID &object_id) {
-  return plasma::plasma_store_runner->IsPlasmaObjectSpillable(object_id);
+  if (plasma::plasma_store_runner != nullptr) {
+    return plasma::plasma_store_runner->IsPlasmaObjectSpillable(object_id);
+  }
+  return false;
 }
 
 void ObjectManager::RunRpcService(int index) {
@@ -865,7 +879,9 @@ void ObjectManager::FillObjectStoreStats(rpc::GetNodeStatsReply *reply) const {
   stats->set_object_store_bytes_used(used_memory_);
   stats->set_object_store_bytes_avail(config_.object_store_memory);
   stats->set_num_local_objects(local_objects_.size());
-  stats->set_consumed_bytes(plasma::plasma_store_runner->GetConsumedBytes());
+  if (plasma::plasma_store_runner) {
+    stats->set_consumed_bytes(plasma::plasma_store_runner->GetConsumedBytes());
+  }
 }
 
 void ObjectManager::Tick(const boost::system::error_code &e) {
@@ -875,13 +891,15 @@ void ObjectManager::Tick(const boost::system::error_code &e) {
 
   // Request the current available memory from the object
   // store.
-  plasma::plasma_store_runner->GetAvailableMemoryAsync([this](size_t available_memory) {
-    main_service_->post(
-        [this, available_memory]() {
-          pull_manager_->UpdatePullsBasedOnAvailableMemory(available_memory);
-        },
-        "ObjectManager.UpdateAvailableMemory");
-  });
+  if (plasma::plasma_store_runner) {
+    plasma::plasma_store_runner->GetAvailableMemoryAsync([this](size_t available_memory) {
+      main_service_->post(
+          [this, available_memory]() {
+            pull_manager_->UpdatePullsBasedOnAvailableMemory(available_memory);
+          },
+          "ObjectManager.UpdateAvailableMemory");
+    });
+  }
 
   pull_manager_->Tick();
